pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "firebreak"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.43.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "starlette>=0.40",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
firebreak-demo = "firebreak.demo:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
